fix - Tipagens de Classes

+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/controller/UsuarioController.kt
@@ -20,19 +20,19 @@ class UsuarioController(
         return ResponseEntity.status(200).body(cadastro)
  }
     @GetMapping("/obterUsuario")
-    fun obterBeneficiarioPorId(id: Int): ResponseEntity<Usuario> {
+    fun obterBeneficiarioPorId(id: Long): ResponseEntity<Usuario> {
         val existeID = beneficiarioService.obterBeneficiarioPorId(id)
         return ResponseEntity.status(200).body(existeID)
     }
 
     @PostMapping("/atualiazarCadastro")
-    fun atualizarBeneficiario(id: Int, novoDTO: BeneficiarioInputDTO): ResponseEntity<BeneficiarioResponseDTO> {
+    fun atualizarBeneficiario(id: Long, novoDTO: BeneficiarioInputDTO): ResponseEntity<BeneficiarioResponseDTO> {
         val attCadastro = beneficiarioService.atualizarBeneficiario(id, novoDTO)
         return ResponseEntity.status(200).body(attCadastro)
     }
 
     @DeleteMapping("/excluirUsuario")
-    fun excluirBeneficiario(id: Int): ResponseEntity<Boolean> {
+    fun excluirBeneficiario(id: Long): ResponseEntity<Boolean> {
         val deletCadastro = beneficiarioService.excluirBeneficiario(id)
         return ResponseEntity.status(200).body(deletCadastro)
     }
